refactor: cards folder, fix style and bugs
